ui(mac): restore polished mobile connection chooser
