+++ Sources/Mobile/Pairing/MobilePairingView.swift
@@ -6,9 +6,9 @@ import SwiftUI
 
 /// The macOS mobile-connection page for pairing an iPhone with this Mac.
 ///
-/// Iroh is the normal account-backed path and is presented first. Tailscale is
-/// an explicit compatibility path that the user opens when a QR or manual
-/// destination is needed.
+/// Iroh is the normal account-backed path and is presented first in the
+/// transport chooser. Tailscale is an explicit compatibility path for QR or
+/// manually-entered destinations.
 struct MobilePairingView: View {
     @State private var model = MobilePairingModel()
     @State private var signInModel = AccountSignInModel(
@@ -20,8 +20,6 @@ struct MobilePairingView: View {
     @State var copiedValue: String?
     /// Bumped per copy so an older flash's dismissal can't clear a newer one.
     @State var copiedValueGeneration = 0
-    /// Whether the optional Tailscale QR flow is currently open.
-    @State private var showingTailscalePairing = false
     /// Reports the scroll content's unconstrained height so the AppKit window
     /// can grow to reveal it while retaining scrolling on shorter displays.
     private let onContentHeightChange: (CGFloat) -> Void
@@ -30,10 +28,6 @@ struct MobilePairingView: View {
     /// when sign-in completes or settles. Captured once; stable post-startup.
     private let coordinator: AuthCoordinator? = AppDelegate.shared?.auth?.coordinator
     private let accountFlow: HostAccountFlow? = AppDelegate.shared?.auth?.accountFlow
-
-    private static let tailscaleDownloadURL = URL(string: "https://tailscale.com/download")!
-    /// Where a Mac user goes to get cmux for iPhone while the beta is invite-only.
-    static let iphoneAppURL = URL(string: "https://github.com/manaflow-ai/cmux#founders-edition")!
 
     init(onContentHeightChange: @escaping (CGFloat) -> Void = { _ in }) {
         self.onContentHeightChange = onContentHeightChange
@@ -62,7 +56,6 @@ struct MobilePairingView: View {
         .onPreferenceChange(MobilePairingContentHeightPreferenceKey.self) { measurement in
             onContentHeightChange(measurement.height)
         }
-        .onAppear { showingTailscalePairing = false }
         .task { await model.refresh() }
         .onDisappear { model.stopObserving() }
         .onChange(of: coordinator?.isAuthenticated ?? false) { _, _ in
@@ -84,47 +77,12 @@ struct MobilePairingView: View {
                 .cmuxFont(.title2, weight: .semibold)
             Text(String(
                 localized: "mobile.connection.subheading",
-                defaultValue: "Sign in to the same cmux account on your iPhone to find this Mac automatically."
+                defaultValue: "Sign in to the same cmux account on your iPhone to find this Mac automatically over Iroh."
             ))
                 .cmuxFont(.callout)
                 .foregroundStyle(.secondary)
                 .fixedSize(horizontal: false, vertical: true)
         }
-    }
-
-    /// The App Store-badge-styled button for getting cmux on the iPhone.
-    private var getIPhoneAppBadge: some View {
-        Link(destination: Self.iphoneAppURL) {
-            HStack(spacing: 8) {
-                Image(systemName: "apple.logo")
-                    .cmuxFont(size: 20)
-                VStack(alignment: .leading, spacing: 0) {
-                    Text(String(
-                        localized: "mobile.pairing.getApp.badge.caption",
-                        defaultValue: "Download cmux for"
-                    ))
-                        .cmuxFont(.caption2)
-                    Text(String(
-                        localized: "mobile.pairing.getApp.badge.platform",
-                        defaultValue: "iPhone"
-                    ))
-                        .cmuxFont(.title3, weight: .semibold)
-                }
-            }
-            .foregroundStyle(.white)
-            .padding(.horizontal, 14)
-            .padding(.vertical, 6)
-            .background(Color.black, in: RoundedRectangle(cornerRadius: 10))
-            .overlay(
-                RoundedRectangle(cornerRadius: 10)
-                    .strokeBorder(Color.white.opacity(0.3))
-            )
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel(String(
-            localized: "mobile.pairing.getApp.link",
-            defaultValue: "Get cmux for iPhone"
-        ))
     }
 
     // MARK: Gated content
@@ -143,11 +101,11 @@ struct MobilePairingView: View {
                     .foregroundStyle(.secondary)
             }
         case let .needsReachableTransport(reachableViaIroh):
-            needsReachableTransportContent(reachableViaIroh: reachableViaIroh)
+            transportContent(.needsReachableTransport(reachableViaIroh: reachableViaIroh))
         case let .failed(message):
             failure(message: message)
         case let .ready(ready):
-            readyContent(ready)
+            transportContent(.ready(ready))
         case .connected:
             connectedContent
         }
@@ -182,289 +140,21 @@ struct MobilePairingView: View {
         .frame(maxWidth: .infinity, minHeight: 200)
     }
 
-    // MARK: Ready
-
-    @ViewBuilder
-    private func readyContent(_ ready: MobilePairingModel.Ready) -> some View {
-        if showingTailscalePairing {
-            VStack(alignment: .center, spacing: 14) {
-                getIPhoneAppBadge
-                tailscaleReadyBody(ready)
-            }
-            .frame(maxWidth: .infinity)
-
-            Divider()
-
-            tailscaleRow(ready)
-            if ready.reachableViaIroh {
-                irohRow(reachableViaIroh: ready.reachableViaIroh)
-            }
-            manualEntry(ready)
-
-            Button {
-                showingTailscalePairing = false
-            } label: {
-                Label(
-                    String(
-                        localized: "mobile.connection.back",
-                        defaultValue: "Back to Mobile Connection"
-                    ),
-                    systemImage: "chevron.left"
-                )
-            }
-            .buttonStyle(.bordered)
-            .controlSize(.small)
-        } else {
-            MobileConnectionOverview(
-                reachableViaIroh: ready.reachableViaIroh,
-                openTailscalePairing: { showingTailscalePairing = true }
-            )
-        }
-
-        footer
-    }
-
-    @ViewBuilder
-    private func tailscaleReadyBody(_ ready: MobilePairingModel.Ready) -> some View {
-        // The spec 4-module quiet zone (white margin) is baked into the QR
-        // bitmap itself, so the code gets no extra white card padding here:
-        // the old 12pt-padded white card doubled the visible quiet zone.
-        // Width is capped so the whole QR, the waiting indicator, and the
-        // transport details all fit the default window without scrolling.
-        MobilePairingQRImageView(payload: ready.attachURL)
-            .frame(maxWidth: 320)
-            .clipShape(RoundedRectangle(cornerRadius: 8))
-            .overlay(
-                RoundedRectangle(cornerRadius: 8)
-                    .strokeBorder(Color.secondary.opacity(0.2))
-            )
-
-        HStack(spacing: 10) {
-            waitingIndicator
-            refreshButton
-        }
-
-        Text(String(
-            localized: "mobile.pairing.scanInstruction",
-            defaultValue: "In cmux on your iPhone, sign in with the same account, choose Tailscale, then scan this code."
-        ))
-        .cmuxFont(.caption)
-        .foregroundStyle(.secondary)
-        .multilineTextAlignment(.center)
-        .fixedSize(horizontal: false, vertical: true)
-
-        if model.availableIOSAppTargets.count > 1 {
-            pairingTargetPicker
-        }
-    }
-
-    private var waitingIndicator: some View {
-        HStack(spacing: 6) {
-            ProgressView().controlSize(.small)
-            Text(String(localized: "mobile.pairing.waiting", defaultValue: "Waiting for your iPhone…"))
-                .cmuxFont(.callout)
-                .foregroundStyle(.secondary)
-        }
-    }
-
-    private var refreshButton: some View {
-        Button(String(localized: "mobile.pairing.refresh", defaultValue: "Refresh Code")) {
-            Task { await model.refresh() }
-        }
-        .buttonStyle(.bordered)
-        .controlSize(.small)
-    }
-
-    private var pairingTargetPicker: some View {
-        HStack(spacing: 6) {
-            Text(String(
-                localized: "mobile.pairing.targetApp",
-                defaultValue: "Open with"
-            ))
-                .cmuxFont(.caption)
-                .foregroundStyle(.secondary)
-            Picker(
-                String(
-                    localized: "mobile.pairing.targetApp",
-                    defaultValue: "Open with"
-                ),
-                selection: Binding(
-                    get: { model.selectedIOSAppTarget },
-                    set: { target in
-                        Task { await model.selectIOSAppTarget(target) }
-                    }
-                )
-            ) {
-                ForEach(model.availableIOSAppTargets) { target in
-                    Text(target.displayName).tag(target)
-                }
-            }
-            .labelsHidden()
-            .pickerStyle(.menu)
-            .controlSize(.small)
-            .fixedSize()
-        }
-    }
-
-    // MARK: No reachable Tailscale route
-
-    @ViewBuilder
-    private func needsReachableTransportContent(reachableViaIroh: Bool) -> some View {
-        if showingTailscalePairing {
-            VStack(alignment: .center, spacing: 14) {
-                getIPhoneAppBadge
-                tailscaleMissingBody
-                Button {
-                    showingTailscalePairing = false
-                } label: {
-                    Label(
-                        String(
-                            localized: "mobile.connection.back",
-                            defaultValue: "Back to Mobile Connection"
-                        ),
-                        systemImage: "chevron.left"
-                    )
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-            }
-            .frame(maxWidth: .infinity)
-        } else {
-            MobileConnectionOverview(
-                reachableViaIroh: reachableViaIroh,
-                openTailscalePairing: { showingTailscalePairing = true }
-            )
-        }
-
-        footer
-    }
-
-    @ViewBuilder
-    private var tailscaleMissingBody: some View {
-        Image(systemName: "network.slash")
-            .cmuxFont(size: 28)
-            .foregroundStyle(.orange)
-        Text(String(
-            localized: "mobile.pairing.req.tailscale.missing",
-            defaultValue: """
-            Tailscale is not connected on this Mac. Install it on both devices \
-            and connect both to the same Tailscale network.
-            """
-        ))
-            .multilineTextAlignment(.center)
-            .foregroundStyle(.secondary)
-            .fixedSize(horizontal: false, vertical: true)
-        Link(
-            String(localized: "mobile.pairing.req.tailscale.get", defaultValue: "Get Tailscale"),
-            destination: Self.tailscaleDownloadURL
+    private func transportContent(_ content: MobilePairingTransportView.Content) -> some View {
+        MobilePairingTransportView(
+            content: content,
+            availableIOSAppTargets: model.availableIOSAppTargets,
+            selectedIOSAppTarget: model.selectedIOSAppTarget,
+            signedInEmail: model.signedInEmail,
+            onRefresh: {
+                Task { await model.refresh() }
+            },
+            onSelectIOSAppTarget: { target in
+                Task { await model.selectIOSAppTarget(target) }
+            },
+            copiedValue: copiedValue,
+            onCopy: flashCopied
         )
-        .buttonStyle(.borderedProminent)
-        refreshButton
-    }
-
-    // MARK: Transport status rows (debugging surface)
-
-    private func irohRow(reachableViaIroh: Bool) -> some View {
-        transportRow(
-            name: "Iroh",
-            healthy: reachableViaIroh,
-            status: reachableViaIroh
-                ? String(localized: "mobile.pairing.transport.status.ready", defaultValue: "Ready")
-                : String(localized: "mobile.pairing.transport.status.notRegistered", defaultValue: "Not registered"),
-            detail: String(
-                localized: "mobile.pairing.transport.iroh.detail",
-                defaultValue: """
-                iPhones signed in to your account find this Mac automatically over Iroh — \
-                end-to-end encrypted, direct when possible, through a cmux relay when not. No code needed.
-                """
-            )
-        ) {
-            EmptyView()
-        }
-    }
-
-    private func tailscaleRow(_ ready: MobilePairingModel.Ready) -> some View {
-        transportRow(
-            name: "Tailscale",
-            healthy: ready.reachableViaTailscale,
-            status: ready.reachableViaTailscale
-                ? String(localized: "mobile.pairing.transport.status.connected", defaultValue: "Connected")
-                : String(localized: "mobile.pairing.transport.status.notDetected", defaultValue: "Not detected"),
-            detail: String(
-                localized: "mobile.pairing.transport.tailscale.detail",
-                defaultValue: "This code pairs over Tailscale instead. Both devices must be connected to the same Tailscale network."
-            )
-        ) {
-            if !ready.reachableViaTailscale {
-                Link(
-                    String(localized: "mobile.pairing.req.tailscale.get", defaultValue: "Get Tailscale"),
-                    destination: Self.tailscaleDownloadURL
-                )
-                .cmuxFont(.caption)
-            }
-        }
-    }
-
-    private func transportRow<Trailing: View>(
-        name: String,
-        healthy: Bool,
-        status: String,
-        detail: String,
-        @ViewBuilder trailing: () -> Trailing
-    ) -> some View {
-        VStack(alignment: .leading, spacing: 2) {
-            HStack(spacing: 6) {
-                Circle()
-                    .fill(healthy ? Color.green : Color.orange)
-                    .frame(width: 6, height: 6)
-                Text(name).cmuxFont(.callout, weight: .medium)
-                Text(status).cmuxFont(.caption).foregroundStyle(.secondary)
-                Spacer(minLength: 8)
-                trailing()
-            }
-            Text(detail)
-                .cmuxFont(.caption)
-                .foregroundStyle(.secondary)
-                .fixedSize(horizontal: false, vertical: true)
-                .padding(.leading, 12)
-        }
-    }
-
-    @ViewBuilder
-    private func manualEntry(_ ready: MobilePairingModel.Ready) -> some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text(String(localized: "mobile.connection.tailscale.manualTitle", defaultValue: "Can't scan? Enter this Mac's Tailscale IP and port:"))
-                .cmuxFont(.caption, weight: .semibold)
-                .foregroundStyle(.secondary)
-            ForEach(ready.tailscaleLines, id: \.self) { line in
-                Text(line).cmuxFont(.caption, design: .monospaced)
-                    .textSelection(.enabled).foregroundStyle(.secondary)
-            }
-            if let entry = ready.manualEntry {
-                HStack(spacing: 8) {
-                    copyButton(label: String(localized: "mobile.pairing.manual.copyIP", defaultValue: "Copy IP"), value: entry.host)
-                    copyButton(label: String(localized: "mobile.pairing.manual.copyPort", defaultValue: "Copy Port"), value: String(entry.port))
-                }
-                .padding(.top, 2)
-            }
-        }
-        .padding(.leading, 12)
-    }
-
-    private var footer: some View {
-        HStack(alignment: .firstTextBaseline) {
-            if let email = model.signedInEmail {
-                Text(String(
-                    format: String(localized: "mobile.pairing.signedInAs", defaultValue: "Signed in as %@"),
-                    locale: .current,
-                    email
-                ))
-                    .cmuxFont(.caption)
-                    .foregroundStyle(.secondary)
-                    .textSelection(.enabled)
-            }
-            Spacer()
-        }
     }
 
 }
